+++ examples/apichange_example.py
@@ -12,7 +12,6 @@
 #!/usr/bin/env python
 
 import time, sys
-
 import wradlib.util as util
 
 
@@ -21,63 +20,105 @@ def help_function(x):
 
 
 def ex_apichange():
-    @util.apichange("0.6.0", par='x', msg="x will take int data in the future")
-    def test_function1(z, x=None):
-        print(x, type(x))
+    @util.apichange_kwarg("0.6.0", par='x', typ=str, msg="x will take int data in the future")
+    def futurechange(z, x=None):
+        if isinstance(x, str):
+            print(z, x, type(x), "normal function behaviour, DeprecationWarning is issued")
+        elif isinstance(x, type(None)):
+            print(z, x, type(x), "normal function behaviour, no DeprecationWarning")
+        else:
+            print(z, x, type(x), "using wrong type here, no DeprecationWarning, "
+                              "but TypeError will be raised")
+            raise TypeError("Wrong Input %s, 'str' expected" % type(x))
         sys.stdout.flush()
         return z, x
 
-    @util.apichange("0.6.0", par='x', typ=str, exfunc=help_function)
-    def test_function2(z, x=None):
-        print(x, type(x))
+    futurechange(0)
+    time.sleep(1)
+    futurechange(1, x='10')
+    time.sleep(1)
+    try:
+        futurechange(2, x=20)
+        time.sleep(1)
+    except TypeError as e:
+        print "Type error: {0}".format(e)
+
+
+    @util.apichange_kwarg("0.6.0", par='x', typ=str, exfunc=help_function)
+    def typechanged(z, x=None):
+        if isinstance(x, int):
+            print(z, x, type(x), "normal function behaviour or type change, "
+                              "DeprecationWarning is issued when 'x' is type(str)")
+        elif isinstance(x, type(None)):
+            print(z, x, type(x), "normal function behaviour, no DeprecationWarning")
+        else:
+            print(z, x, type(x), "using wrong type here, TypeError will be raised")
+            raise TypeError("Wrong Input %s, 'int' expected" % type(x))
         sys.stdout.flush()
         return z, x
 
-    @util.apichange("0.6.0", par='x', expar='y', typ=str, exfunc=help_function)
-    def test_function3(z, y=None):
-        print(y, type(y))
+    print("Test typechange_kwarg")
+    typechanged(0)
+    time.sleep(1)
+    typechanged(3, x='30')
+    time.sleep(1)
+    typechanged(4, x=40)
+    time.sleep(1)
+
+
+    @util.apichange_kwarg("0.6.0", par='x', typ=str, expar='y')
+    def namechanged(z, y=None):
+        if isinstance(y, str):
+            print(z, y, type(y), "DeprecationWarning")
+        elif isinstance(y, type(None)):
+            print(z, y, type(y), "normal function behaviour, no DeprecationWarning")
+        else:
+            print(z, y, type(y), "using wrong type here, TypeError is issued")
+            raise TypeError("Wrong Input %s, 'str' expected" % type(y))
         sys.stdout.flush()
         return z, y
 
-    # this will run normal, because x is no parameter
-    test_function1(1)
-    sys.stderr.flush()
+    print("Test namechange_kwarg")
+    namechanged(0)
+    time.sleep(1)
+    namechanged(5, x='50')
+    time.sleep(1)
+    try:
+        namechanged(6, x=60)
+    except TypeError as e:
+        print "Type error: {0}".format(e)
+    time.sleep(1)
+    namechanged(7, y='70')
+    time.sleep(1)
+    try:
+        namechanged(8, y=80)
+    except TypeError as e:
+        print "Type error: {0}".format(e)
     time.sleep(1)
 
-    # this will display the warning message, because x is parameter
-    test_function1(1, x='10')
-    sys.stderr.flush()
+    @util.apichange_kwarg("0.6.0", par='x', typ=str, expar='y', exfunc=help_function)
+    def name_and_type_changed(z, y=None):
+        if isinstance(y, int):
+            print(z, y, type(y), "normal function behaviour or paramter and type change, "
+                              "DeprecationWarning is issued when 'x' is given")
+        elif isinstance(y, type(None)):
+            print(z, y, type(y), "normal function behaviour, no DeprecationWarning")
+        else:
+            print(z, y, type(y), "using wrong type here, TypeError will be raised")
+            raise TypeError("Wrong Input %s, 'str' expected" % type(y))
+        return z, y
+
+    print("Test apichange_kwarg type and name")
+    name_and_type_changed(0)
+    time.sleep(1)
+    name_and_type_changed(9, x='90')
+    time.sleep(1)
+    try:
+        name_and_type_changed(10, x=100)
+    except TypeError as e:
+        print "Type error: {0}".format(e)
     time.sleep(1)
 
-    # this will display the warning message, because x is parameter
-    test_function1(1, x=20)
-    sys.stderr.flush()
-    time.sleep(1)
-
-    # this will display the warning message, because x is string,
-    # use the help_function to convert to int type
-    test_function2(1, x='30')
-    sys.stderr.flush()
-    time.sleep(1)
-
-    # this will run normal, because x is no string parameter
-    test_function2(1, x=40)
-    sys.stderr.flush()
-    time.sleep(1)
-
-
-    # this will display the warning message, because x is string
-    # use the help_function to convert to int type
-    # switch par (x) to expar (y)
-    test_function3(1, x='50')
-    sys.stderr.flush()
-    time.sleep(1)
-
-    # this will display the warning message, because x is parameter
-    # and expar is given
-    # switch par (x) to expar (y)
-    test_function3(1, x=60)
-    sys.stderr.flush()
 
 if __name__ == '__main__':
     ex_apichange()