+++ wradlib/util.py
@@ -37,7 +37,8 @@ import importlib
 
 import warnings
 import functools
-warnings.simplefilter('once', DeprecationWarning)
+warnings.simplefilter('always', DeprecationWarning)
+#warnings.simplefilter('always', FutureWarning)
 
 def deprecated(replacement=None):
     """A decorator which can be used to mark functions as deprecated.
@@ -81,8 +82,8 @@ def deprecated(replacement=None):
     return outer
 
 
-def apichange(ver, par, typ=None, msg=None, expar=None, exfunc=None):
-    """A decorator to generate an api-change warning.
+def apichange_kwarg(ver, par, typ, expar=None, exfunc=None, msg=None):
+    """A decorator to generate a DeprectationWarning.
 
     .. versionadded:: 0.5.0
 
@@ -100,70 +101,39 @@ def apichange(ver, par, typ=None, msg=None, expar=None, exfunc=None):
         Name of parameter which is affected
     typ : Python type
         Data type of parameter which is affected
-    msg : string
-        additional warning message
     expar : string
         Name of parameter to be used in future
     exfunc : function
         Function which can convert from old to new parameter
-
-    Examples:
-    ---------
-    Parameter Type change:
-    >>> # plot the deprecation message
-    >>> @apichange("0.6.0", par='proj', msg="'proj' will change from proj4-string to osr-object")
-    >>> def plot_ppi(data, r=None, az=None, autoext=True,
-    ...              site=(0,0), proj=None, elev=0.,
-    ...              ax=None, **kwargs):
-
-    >>> # if the api already changed, we can issue warning and try to convert
-    >>> @apichange("0.6.0", par='proj', typ=str, exfunc=georef.proj4_to_osr)
-    >>> def plot_ppi(data, r=None, az=None, autoext=True,
-    ...              site=(0,0), proj=None, elev=0.,
-    ...              ax=None, **kwargs):
-
-    One Parameter replaced by another
-    >>> # plot the deprecation message
-    >>> @apichange("0.6.0", par='projstr', expar='proj')
-    >>> def volcoords_from_polar(sitecoords, elevs, azimuths, ranges, projstr=None):
-
-    >>> # if the api already changed, we can issue warning and try to convert
-    >>> @apichange("0.6.0", par='projstr', typ=str, expar='proj', exfunc=georef.proj4_to_osr)
-    >>> def volcoords_from_polar(sitecoords, elevs, azimuths, ranges, proj=None):
+    msg : string
+        additional warning message
 
     """
     def outer(func):
-        wmsg = "\nPrevious behaviour of parameter '%s' in function '%s.%s' is deprecated " \
-               "\nand will be changed in version '%s'." % (par, func.__module__,func.__name__, ver)
-        if exfunc:
-            wmsg += "\nWrong parameter types will be automatically converted by " \
-            "using %s.%s." % (exfunc.__module__, exfunc.func_name)
-        if expar is not None:
-            wmsg += "\nUse parameter %s instead." % expar
-        if msg is not None:
-            wmsg += "\n%s" % msg
-
         @functools.wraps(func)
         def inner(*args, **kwargs):
-            para = kwargs.get(par, None)
+            para = kwargs.pop(par, None)
+            key = par
             if para:
-                if typ:
-                    if isinstance(para, typ):
-                        warnings.warn(wmsg, category=FutureWarning, stacklevel=2)
-                        if exfunc:
-                            para = exfunc(para)
+                wmsg = "\nPrevious behaviour of parameter '%s' in function '%s.%s' is deprecated " \
+                       "\nand will be changed in version '%s'." % (par, func.__module__,func.__name__, ver)
+                if expar:
+                    wmsg += "\nUse parameter %s instead." % expar
+                if exfunc:
+                    wmsg += "\nWrong parameter types will be automatically converted by " \
+                            "using %s.%s." % (exfunc.__module__, exfunc.func_name)
+                if msg:
+                    wmsg += "\n%s" % msg
+                if isinstance(para, typ):
+                    if exfunc:
+                        para = exfunc(para)
                     if expar:
-                        warnings.warn(wmsg, category=FutureWarning, stacklevel=2)
-                        kwargs.pop(par)
-                        kwargs.update({expar: para})
-                    else:
-                        kwargs.update({par: para})
-                else:
-                    warnings.warn(wmsg, category=FutureWarning, stacklevel=2)
+                        key = expar
+                    warnings.warn(wmsg, category=DeprecationWarning, stacklevel=2)
+                kwargs.update({key: para})
             return func(*args, **kwargs)
         return inner
     return outer
-
 
 class OptionalModuleStub(object):
     """Stub class for optional imports.